pyproject.toml
[project]
name = "qbt-rss-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "python-dotenv>=1.2.1",
    "qbittorrent-api>=2025.11.1",
]
